move to the simple files login treatment plan routes
