+++ src/routes/login.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 
-import { login } from '@controllers/login'
+import { login } from '@controllers/Login/login'
 
 const loginRouter = Router()
